+++ test/libsolidity/smtCheckerTests/functions/function_external_call_should_not_inline_1.sol
@@ -13,3 +13,4 @@ contract C {
 	}
 }
 // ----
+// Warning 6321: (85-89): Unnamed return variable can remain unassigned. Add an explicit return with value to all non-reverting code paths or name the variable.