+++ test/libsolidity/smtCheckerTests/functions/internal_call_state_var_init_2.sol
@@ -7,3 +7,4 @@ contract c {
 	bool b = (f() > 0) || (f() > 0);
 }
 // ----
+// Warning 6321: (86-90): Unnamed return variable can remain unassigned. Add an explicit return with value to all non-reverting code paths or name the variable.