+++ test/libsolidity/smtCheckerTests/types/rational_large_1.sol
@@ -7,4 +7,5 @@ contract c {
 	}
 }
 // ----
+// Warning 6321: (80-84): Unnamed return variable can remain unassigned. Add an explicit return with value to all non-reverting code paths or name the variable.
 // Warning 6328: (128-142): CHC: Assertion violation happens here.