+++ Changelog.md
@@ -22,6 +22,8 @@ Language Features:
 
 Compiler Features:
  * SMTChecker: Support inline arrays.
+ * Control Flow Graph: Print warning for non-empty functions with unnamed return parameters that are not assigned a value in all code paths.
+
 
 Bugfixes:
  * Code generator: Fix internal compiler error when referencing members via module name but not using the reference.